Extracting resource, some naming changes.
