+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/indexer/responses/IndexSetFieldTypeSummary.java
@@ -25,10 +25,8 @@ public record IndexSetFieldTypeSummary(@JsonProperty(INDEX_SET_ID) String id,
                                        @JsonProperty(STREAM_TITLES) List<String> streams,
                                        @JsonProperty(FIELD_TYPE_HISTORY) List<String> fieldTypeHistory
 ) {
-
-
     public static final String INDEX_SET_ID = "index_set_id";
     public static final String INDEX_SET_TITLE = "index_set_title";
     public static final String STREAM_TITLES = "stream_titles";
-    public static final String FIELD_TYPE_HISTORY = "field_type_history";
+    public static final String FIELD_TYPE_HISTORY = "types";
 }
